begin taak selectie

+++ src/main/java/com/roostermaker/Kind.java
@@ -2,7 +2,7 @@ package com.roostermaker;
 public class Kind extends Gezin {
 
 
-    public Kind(String naam, double beschikbareTijd, String gezinsverhouding) {
+    public Kind(String naam, int beschikbareTijd, String gezinsverhouding) {
         super(naam, beschikbareTijd, gezinsverhouding);
     }
