+++ src/main/java/com/roostermaker/Ouder.java
@@ -2,7 +2,7 @@ package com.roostermaker;
 
 public class Ouder extends Gezin {
     
-    public Ouder(String naam, double beschikbareTijd, String gezinsverhouding) {
+    public Ouder(String naam, int beschikbareTijd, String gezinsverhouding) {
         super(naam, beschikbareTijd, gezinsverhouding);
     }
